requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
numpy==1.16.2
pandas==0.24.1
python-crontab==2.3.6
python-dateutil==2.7.3
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.0
urllib3==1.24.1
